- ported hr_webcam to 8.0 - updated __openerp__.py and hr_webcam_view.xml to reflect changes to 8.0 api - modified hr_webcam.css (wider camera viewport)
