+++ hr_webcam/__openerp__.py
@@ -1,12 +1,12 @@
-# -*- coding:utf-8 -*-
+#-*- coding:utf-8 -*-
 #
 #
 #    Copyright (C) 2013 Michael Telahun Makonnen <mmakonnen@gmail.com>.
 #    All Rights Reserved.
 #
 #    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as published
-#    by the Free Software Foundation, either version 3 of the License, or
+#    it under the terms of the GNU Affero General Public License as published by
+#    the Free Software Foundation, either version 3 of the License, or
 #    (at your option) any later version.
 #
 #    This program is distributed in the hope that it will be useful,
@@ -50,7 +50,6 @@ Capture employee pictures with an attached web cam.
         'hr_webcam_data.xml',
         'hr_webcam_view.xml',
     ],
-    'test': [
-    ],
-    'installable': False,
+    'installable': True,
+    'active': False,
 }
