+++ hr_webcam/hr_webcam_view.xml
@@ -1,6 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <openerp>
     <data>
+        <template id="assets_backend" name="hr_webcam assets" inherit_id="web.assets_backend">
+            <xpath expr="." position="inside">
+                <link rel="stylesheet" href="/hr_webcam/static/src/css/hr_webcam.css"/>
+                <script type="text/javascript" src="/hr_webcam/static/src/js/jquery.webcam.js"></script>
+                <script type="text/javascript" src="/hr_webcam/static/src/js/hr_webcam.js"></script>
+            </xpath>
+        </template>
         
         <record id="action_take_photo" model="ir.actions.client">
             <field name="name">Take Picture</field>
