+++ hr_webcam/static/src/js/hr_webcam.js
