+++ hr_webcam/static/src/js/jquery.webcam.js
