+++ hr_webcam/static/src/xml/hr_webcam.xml
